Move tests for matchers into a separate file
The test_responses file is getting big, and this separate mirrors how
the code is separated. I didn't change the tests as I think having
integration tests are important as that's how users use responses.
